Add generic conversions to currency trait
Here weadd to<T> and from<T> for converting to or from other currencies.
I believe the from method will actually be more useful as a general
rule, but I believe it'll also look weird to most programmers, because
it is primarily intended to take advantage of Rust's type inference,
which is not a thing unless you're writing rust.

In the progrma, we provide an exemplar of to<T> in action.